feat: move some values to `builder` module
