+++ backend/backend/settings/worker/events/common.py
@@ -1,3 +1,4 @@
+from ...deps.image_build import *
 from ...deps.ledger import *
 from ...deps.orchestrator import *
 from ...dev import *
