+++ backend/builder/tests/test_task_build_image.py
@@ -1,0 +1,17 @@
+import pytest
+
+from builder.exceptions import BuildError
+from substrapp.tasks import tasks_compute_task
+
+
+@pytest.mark.django_db
+def test_store_failure_build_error():
+    compute_task_key = "42ff54eb-f4de-43b2-a1a0-a9f4c5f4737f"
+    msg = "Error building image"
+    exc = BuildError(msg)
+
+    failure_report = tasks_compute_task._store_failure(exc, compute_task_key)
+    failure_report.refresh_from_db()
+
+    assert str(failure_report.compute_task_key) == compute_task_key
+    assert failure_report.logs.read() == str.encode(msg)
