+++ backend/builder/volumes.py
@@ -1,0 +1,7 @@
+import os
+
+from django.conf import settings
+
+
+def get_docker_cache_pvc_name():
+    return f"{settings.WORKER_PVC_DOCKER_CACHE}-{os.getenv('HOSTNAME')}"
