+++ backend/substrapp/kubernetes_utils.py
@@ -1,5 +1,4 @@
 import enum
-import time
 
 import kubernetes
 import structlog
@@ -7,15 +6,11 @@ from django.conf import settings
 
 from substrapp.exceptions import KubernetesError
 from substrapp.exceptions import PodDeletedError
-from substrapp.exceptions import PodError
 from substrapp.exceptions import PodReadinessTimeoutError
-from substrapp.exceptions import PodTimeoutError
-from substrapp.utils import timeit
 
 logger = structlog.get_logger(__name__)
 
 NAMESPACE = settings.NAMESPACE
-HTTP_CLIENT_TIMEOUT_SECONDS = settings.HTTP_CLIENT_TIMEOUT_SECONDS
 RUN_AS_GROUP = settings.COMPUTE_POD_RUN_AS_GROUP
 RUN_AS_USER = settings.COMPUTE_POD_RUN_AS_USER
 FS_GROUP = settings.COMPUTE_POD_FS_GROUP
@@ -63,159 +58,6 @@ def get_security_context(root: bool = False, capabilities: list[str] = None) -> 
     return security_context
 
 
-class PodState:
-    def __init__(self, status: ObjectState, reason: str = "", message: str = ""):
-        self.status = status
-        self.reason = reason
-        self.message = message
-
-    def set_reason(self, container_status: kubernetes.client.V1ContainerState) -> None:
-        if self.status == ObjectState.WAITING:
-            self.reason = container_status.waiting.reason
-            self.message = container_status.waiting.message
-        if self.status == ObjectState.FAILED:
-            self.reason = container_status.terminated.reason
-            self.message = container_status.terminated.message
-
-
-def watch_pod(k8s_client: kubernetes.client.CoreV1Api, name: str):
-    """Watch a Kubernetes pod status
-    It will observe all the containers inside the pod and return when the pod will
-    reach the Completed state. If the pod is pending indefinitely or fail, an exception will be raised.
-    Args:
-        k8s_client (kubernetes.client.CoreV1Api): Kubernetes API client
-        name (str): name of the pod to watch
-    Raises:
-        PodError: this exception is raised if the pod exits with an error
-        PodTimeoutError: this exception is raised if the pod does not reach the running state after some time
-    """
-    attempt = 0
-    # with 60 attempts we wait max 2 min with a pending pod
-    max_attempts = 60
-
-    # This variable is used to track the current status through retries
-    previous_pod_status = None
-
-    while attempt < max_attempts:
-        try:
-            api_response = retrieve_pod_status(k8s_client, name)
-        except kubernetes.client.ApiException as exc:
-            logger.warning("Could not retrieve pod status", pod_name=name, exc_info=exc)
-            attempt += 1
-            time.sleep(0.2)
-            continue
-
-        pod_state = _get_pod_state(api_response)
-
-        if pod_state.status != previous_pod_status:
-            previous_pod_status = pod_state.status
-            logger.info(
-                "Pod status changed",
-                pod_name=name,
-                status=pod_state.status,
-                reason=pod_state.reason,
-                message=pod_state.message,
-                attempt=attempt,
-                max_attempts=max_attempts,
-            )
-
-        if pod_state.status == ObjectState.COMPLETED:
-            return
-
-        if pod_state.status == ObjectState.FAILED:
-            raise PodError(f"Pod {name} terminated with error: {pod_state.reason}")
-
-        if pod_state.status == ObjectState.PENDING:
-            # Here we basically consume a free retry everytime but we still need to
-            # increment attempt because if at some point our pod is stuck in pending state
-            # we need to exit this function
-            attempt += 1
-            time.sleep(2)
-
-        # Here PodInitializing and ContainerCreating are valid reasons to wait more time
-        # Other possible reasons include "CrashLoopBackOff", "CreateContainerConfigError",
-        # "ErrImagePull", "ImagePullBackOff", "CreateContainerError", "InvalidImageName"
-        if (
-            pod_state.status == ObjectState.WAITING
-            and pod_state.reason not in ["PodInitializing", "ContainerCreating"]
-            or pod_state.status == ObjectState.UNKNOWN
-        ):
-            attempt += 1
-
-        time.sleep(0.2)
-
-    raise PodTimeoutError(f"Pod {name} didn't complete after {max_attempts} attempts")
-
-
-def _get_pod_state(pod_status: kubernetes.client.V1PodStatus) -> PodState:
-    """extracts the current pod state from the PodStatus Kubernetes object
-    Args:
-        pod_status (kubernetes.client.models.V1PodStatus): A Kubernetes PodStatus object
-    """
-    if pod_status.phase in ["Pending"]:
-        # On the first query the pod just created and often pending as it is not already scheduled to a node
-        return PodState(ObjectState.PENDING, pod_status.reason, pod_status.message)
-
-    container_statuses: list[kubernetes.client.V1ContainerStatus] = (
-        pod_status.init_container_statuses if pod_status.init_container_statuses else []
-    )
-    container_statuses += pod_status.container_statuses
-
-    completed_containers = 0
-    for container in container_statuses:
-        container_state: ObjectState = _get_container_state(container)
-
-        if container_state in [ObjectState.RUNNING, ObjectState.WAITING, ObjectState.FAILED]:
-            pod_state = PodState(container_state)
-            pod_state.set_reason(container.state)
-            return pod_state
-        if container_state == ObjectState.COMPLETED:
-            completed_containers += 1
-
-    if completed_containers == len(container_statuses):
-        return PodState(ObjectState.COMPLETED, "", "pod successfully completed")
-
-    logger.debug("pod status", pod_status=pod_status)
-    return PodState(ObjectState.UNKNOWN, "", "Could not deduce the pod state from container statuses")
-
-
-def _get_container_state(container_status: kubernetes.client.V1ContainerStatus) -> ObjectState:
-    """Extracts the container state from a ContainerStatus Kubernetes object
-    Args:
-        container_status (kubernetes.client.models.V1ContainerStatus): A ContainerStatus object
-    Returns:
-        ObjectState: the state of the container
-    """
-    # Here we need to check if we are in a failed state first since kubernetes will retry
-    # we can end up running after a failure
-    if container_status.state.terminated:
-        if container_status.state.terminated.exit_code != 0:
-            return ObjectState.FAILED
-        else:
-            return ObjectState.COMPLETED
-    if container_status.state.running:
-        return ObjectState.RUNNING
-    if container_status.state.waiting:
-        return ObjectState.WAITING
-    return ObjectState.UNKNOWN
-
-
-def pod_exists(k8s_client, name: str) -> bool:
-    try:
-        k8s_client.read_namespaced_pod(name=name, namespace=NAMESPACE)
-    except kubernetes.client.ApiException:
-        return False
-    else:
-        return True
-
-
-def retrieve_pod_status(k8s_client: kubernetes.client.CoreV1Api, pod_name: str) -> kubernetes.client.V1PodStatus:
-    pod: kubernetes.client.V1Pod = k8s_client.read_namespaced_pod_status(
-        name=pod_name, namespace=NAMESPACE, pretty=True
-    )
-    return pod.status
-
-
 def pod_exists_by_label_selector(k8s_client: kubernetes.client.CoreV1Api, label_selector: str) -> bool:
     """Return True if the pod exists, else False.
 
@@ -229,18 +71,6 @@ def pod_exists_by_label_selector(k8s_client: kubernetes.client.CoreV1Api, label_
     """
     res = k8s_client.list_namespaced_pod(namespace=NAMESPACE, label_selector=label_selector)
     return len(res.items) > 0
-
-
-@timeit
-def get_pod_logs(k8s_client, name: str, container: str, ignore_pod_not_found: bool = False) -> str:
-    try:
-        return k8s_client.read_namespaced_pod_log(name=name, namespace=NAMESPACE, container=container)
-    except kubernetes.client.ApiException as exc:
-        if ignore_pod_not_found and exc.reason == "Not Found":
-            return f"Pod not found: {NAMESPACE}/{name} ({container})"
-        if exc.reason == "Bad Request":
-            return f"In {NAMESPACE}/{name} \n {str(exc.body)}"
-        return f"Unable to get logs for pod {NAMESPACE}/{name} ({container}) \n {str(exc)}"
 
 
 def delete_pod(k8s_client, name: str) -> None:
